Function 6 and 7
Function 6: Tokenization
It is the process of splitting up a large body of text into smaller lines or words. In this function, we are using a Twitter data frame to access tweets arranged in a series and split those tweets into smaller words.

Function 7: Stop Words
The use of the function 6 to remove stop words from the individual tweets.